requirements.txt
GitPython==3.1.40
matplotlib==3.8.2
tabulate==0.9.0